The new "I Love You" virus is not the work of some snot-nosed acne-laced teenager working from a basement in the Phillipines. It's actually part of a conspiracy concocted by the unholy alliance of Microsoft and several well-known and well-despised spammers.
You'll notice that the ILOVEYOU, Melissa, and Tuxissa strains all extract
email addresses from the victim's system. This is a gold mine for
spammers, who are able to use these viruses to harvest active email
addresses for them. Everytime ILOVEYOU, for instance, propogates, it keeps
track of all the email addresses it has been sent to, so that when it
finally boomerangs back to a spammer, they have a nice convenient list of
addresses to send "laser printer toner" and "get rich quick!"
advertisements to.

   -- Bob Smith (not his real code-name), in a speech given at the
      First Annual Connecticut Conspiracy Convention (ConConCon),
      "the largest ever gathering of conspiracy theorists east of the
      Mississippi."